Move our DSL to its own class
