+++ lib/wcc/contentful_model/menu.rb
@@ -3,5 +3,5 @@
 class WCC::ContentfulModel::Menu < WCC::ContentfulModel
   validate_field :name, :String
   validate_field :icon, :Asset, :optional
-  validate_field :first_group, :Link, :array, link_to: %w[menu menuItem]
+  validate_field :first_group, :Array, link_to: %w[menu menuItem]
 end
