+++ spec/wcc/contentful/model_validators_spec.rb
@@ -1,5 +1,3 @@
-
-
 # frozen_string_literal: true
 
 RSpec.describe(WCC::Contentful::ModelValidators) do
